+++ app/src/main/java/org/schabi/newpipe/Downloader.java
@@ -30,7 +30,7 @@ import info.guardianproject.netcipher.NetCipher;
  * along with NewPipe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-public class Downloader implements org.schabi.newpipe.services.Downloader {
+public class Downloader implements org.schabi.newpipe.crawler.Downloader {
     
     private static final String USER_AGENT = "Mozilla/5.0 (Windows NT 6.1; WOW64; rv:43.0) Gecko/20100101 Firefox/43.0";
