+++ app/src/main/java/org/schabi/newpipe/crawler/StreamingService.java
@@ -1,15 +1,10 @@
-package org.schabi.newpipe.services;
-
-import org.schabi.newpipe.VideoPreviewInfo;
-
-import java.util.ArrayList;
-import java.util.Vector;
+package org.schabi.newpipe.crawler;
 
 /**
- * Created by Christian Schabesberger on 10.08.15.
+ * Created by Christian Schabesberger on 23.08.15.
  *
  * Copyright (C) Christian Schabesberger 2015 <chris.schabesberger@mailbox.org>
- * SearchEngine.java is part of NewPipe.
+ * StreamingService.java is part of NewPipe.
  *
  * NewPipe is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,16 +20,16 @@ import java.util.Vector;
  * along with NewPipe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-@SuppressWarnings("ALL")
-public interface SearchEngine {
-    class Result {
-        public String errorMessage = "";
-        public String suggestion = "";
-        public final Vector<VideoPreviewInfo> resultList = new Vector<>();
+public interface StreamingService {
+    class ServiceInfo {
+        public String name = "";
     }
+    ServiceInfo getServiceInfo();
+    VideoExtractor getExtractorInstance(String url, Downloader downloader);
+    SearchEngine getSearchEngineInstance();
 
-    ArrayList<String> suggestionList(String query, Downloader dl);
-
-    //Result search(String query, int page);
-    Result search(String query, int page, String contentCountry, Downloader dl);
+    /**When a VIEW_ACTION is caught this function will test if the url delivered within the calling
+    Intent was meant to be watched with this Service.
+    Return false if this service shall not allow to be called through ACTIONs.*/
+    boolean acceptUrl(String videoUrl);
 }
