+++ app/src/main/java/org/schabi/newpipe/crawler/VideoExtractor.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.services;
+package org.schabi.newpipe.crawler;
 
 /**
  * Created by Christian Schabesberger on 10.08.15.
