+++ app/src/main/java/org/schabi/newpipe/crawler/VideoPreviewInfo.java
@@ -1,10 +1,10 @@
-package org.schabi.newpipe;
+package org.schabi.newpipe.crawler;
 
 import android.graphics.Bitmap;
 import android.os.Parcel;
 import android.os.Parcelable;
 
-import org.schabi.newpipe.services.AbstractVideoInfo;
+import org.schabi.newpipe.crawler.AbstractVideoInfo;
 
 /**
  * Created by Christian Schabesberger on 26.08.15.
